[registry-facade]: export config as package
