+++ components/registry-facade/cmd/root.go
@@ -5,7 +5,6 @@
 package cmd
 
 import (
-	"encoding/json"
 	"fmt"
 	"os"
 
@@ -13,7 +12,6 @@ import (
 
 	"github.com/gitpod-io/gitpod/common-go/log"
 	"github.com/gitpod-io/gitpod/common-go/tracing"
-	"github.com/gitpod-io/gitpod/registry-facade/pkg/registry"
 )
 
 var (
@@ -49,28 +47,4 @@ func Execute() {
 func init() {
 	rootCmd.PersistentFlags().BoolVarP(&jsonLog, "json-log", "j", true, "produce JSON log output on verbose level")
 	rootCmd.PersistentFlags().BoolVarP(&verbose, "verbose", "v", false, "Enable verbose JSON logging")
-}
-
-// Config configures this servuce
-type Config struct {
-	Registry       registry.Config `json:"registry"`
-	AuthCfg        string          `json:"dockerAuth"`
-	PProfAddr      string          `json:"pprofAddr"`
-	PrometheusAddr string          `json:"prometheusAddr"`
-}
-
-// getConfig loads and validates the configuration
-func getConfig(fn string) (*Config, error) {
-	fc, err := os.ReadFile(fn)
-	if err != nil {
-		return nil, err
-	}
-
-	var cfg Config
-	err = json.Unmarshal(fc, &cfg)
-	if err != nil {
-		return nil, err
-	}
-
-	return &cfg, nil
 }
